Fix incorrect arena overflow check

+++ ir_private.h
@@ -257,7 +257,7 @@ IR_ALWAYS_INLINE void* ir_arena_alloc(ir_arena **arena_ptr, size_t size)
 	ir_arena *arena = *arena_ptr;
 	char *ptr = (char*)IR_ALIGNED_SIZE((uintptr_t)arena->ptr, 8);
 
-	if (EXPECTED(size <= (size_t)(arena->end - ptr))) {
+	if (EXPECTED((ssize_t)size <= (ssize_t)(arena->end - ptr))) {
 		arena->ptr = ptr + size;
 	} else {
 		size_t arena_size =
